fix: Simplify save_figures
Remove unnecessary type checking and casting code for the output path
parameter; this is an internal function and the caller should ensure
that the path is a a proper pathlib.Path (which your type hint says it
is anyway). Simplify the directory creation code following the same
approach found in the rest of the codebase.

Minor:
- Rename function to save_figure
- Drop new line between docstring and code

Signed-off-by: Phoevos Kalemkeris <phoevos@protonmail.com>