fix: Handle error message on chat with stream (#31)
Properly handle received errors when streaming. Previously, whenever
an error was returned by the backend, the frontend would fail with a
'TypeError: Cannot read properties of undefined (reading "content")'
while attempting to retrieve 'event["delta"]["content"]', resulting in the
following if-cases never running. Adding a check for the existence of
'event["delta"]' successfully mitigates the issue.

Signed-off-by: nextedoff <rajmukx22@gmail.com>